[ScriptQuicklook] Corrected JSON metadata processing

+++ script_quicklook_new_temp.py
@@ -34,7 +34,6 @@ if dir_out is None:
 preserve_original = True
 make_rgb = False
 
-json_cover = r'\\172.21.195.2/FTP-Share/ftp/roslesinforg/balaganskoe/S3/SNT_сover.json'#r'\\172.21.195.2/FTP-Share/ftp/images/region82/vector_cover.json'
 vector_granule_path = r'\\172.21.195.2\FTP-Share\ftp\images\granules_grid.shp'
 ms2pms = True
 invert_red_blue = False
@@ -88,7 +87,6 @@ def get_pms_json(path_cover, path_out, pms_id, pms_raster_path=''):
         return 1
 
     ms_id = pms_id.replace('.PMS', '.MS')
-    print(ms_id)
     filter_dataset_by_col(path_cover, 'id', ms_id, path_out=path_out)
 
     pms_ds, pms_lyr = get_lyr_by_path(path_out, 1)
@@ -376,17 +374,21 @@ for i, path_in in enumerate(path_in_list):
                     if lyr_.GetFeatureCount()==1:
                         granule_metadata_json(tpath, tpath_meta, json_out, path_in, ms2pms=ms2pms)
                 elif ms2pms:
-                    print(n)
                     get_pms_json(json_cover, json_out, n, pms_raster_path=path_in)
                 else:
                     filter_dataset_by_col(json_cover, 'id', n, path_out=json_out)
-                ds_out, lyr_out = get_lyr_by_path(json_out)
-                if len(lyr_out)==0:
-                    ds_out = None
-                    os.remove(json_out)
+                if os.path.exists(json_out):
+                    ds_out, lyr_out = get_lyr_by_path(json_out)
+                    if lyr_out is None:
+                        print('JSON metadata file not created: %s' % n)
+                        ds_out = None
+                        os.remove(json_out)
+                    elif len(lyr_out)==0:
+                        print('JSON metadata file is empty: %s' % n)
+                        ds_out = None
+                        os.remove(json_out)
             else:
                 print('%i -- file exists -- %s.json' % (i, n))
-
 
 if report:
     TotalCover(fullpath(dir_out, 'scene_cover.json'), export_data)
